Create new styling for trip plan
