+++ frontend/app/scripts/result/_result.ts
@@ -10,11 +10,6 @@ module Result {
       'app.common',
       'ngDialog'
     ])
-    .constant('RESULT_STATE', {
-      overview: 0,
-      details: 1,
-      hotel: 2
-    })
     .constant('VEHICLE_TYPE', {
       bed: 0,
       street_view: 1,
@@ -31,6 +26,7 @@ module Result {
     .constant('SEGMENT_ZOOM_WIDTH', 40) // size to zoom the segment to
     .constant('TRANSFER_TIME', 5) // how much time is added between trip segments
     .controller('resultCtrl', ResultCtrl)
+    .controller('resultDetailsCtrl', ResultDetailsCtrl)
     .service('tripApi', TripApi)
     .service('tripCache', TripCache)
     .service('itineraryHelper', ItineraryHelper)
