+++ frontend/i18n/en.yaml
@@ -89,21 +89,17 @@ result:
   connection:
     towards: towards
   details:
-    title: Itinerary details
-    tripoverview:
-      title: Trip overview
-      price: Price
-      duration: Duration
-      departuretime: Departure time
-      arrivaltime: Arrival time
+    tripoverview: Trip Overview
+    itinerary: Itinerary Details
     instructions:
-      text: Please check the trip details before proceeding to booking.
-      overnight: Your trip requires an overnight stay. You can easily book the hotel proposal, select a different hotel or deselect the hotel option.
-    back: Back to overview
-    travelplan: Travel Plan
-    hotels: Get available hotels
-  tripsegment:
-    flight: flight
+      overnight: This trip requires an overnight stay. You can easily select from a list of suitable hotels that are located close to your travel route.
+    action:
+      hotels: Open hotel selection
+      back: Back to overview
+      download: Download .ics
+      print: Print
+      booknow: Book now
+      bookinglink: booking link
   hotelspage:
     title: Availabel hotels
     list:
@@ -112,7 +108,6 @@ result:
       price: Price
       url: Hotel url
       distance: Distance from destination
-    back: Go back
 print:
   title: Trip overview
   summary:
