+++ frontend/styles/elements/_colors.scss
@@ -2,7 +2,7 @@
 $c-orange: #ffc233;
 $c-white: #fff;
 $c-grey: #b9b9b9;
-$c-grey-dark: #7a7a7a;
+$c-grey-dark: rgba(122, 122, 122, 0.67);
 $c-green: #5cbd3d;
 $c-red: #d92929;
 $c-creme: #ffefc8;
